refactor: changed namespace and used service provider
