+++ models/classes/import/ServiceProvider/ImportServiceProvider.php
@@ -15,13 +15,14 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  *
- * Copyright (c) 2021 (update and modification) Open Assessment Technologies SA;
+ * Copyright (c) 2021-2023 (update and modification) Open Assessment Technologies SA;
  */
 
 declare(strict_types=1);
 
 namespace oat\tao\model\import\ServiceProvider;
 
+use oat\tao\model\import\CustomizedRdfImporter;
 use oat\tao\model\upload\UploadService;
 use oat\tao\model\import\service\AgnosticImportHandler;
 use oat\tao\model\StatisticalMetadata\Import\Processor\ImportProcessor;
@@ -65,5 +66,9 @@ class ImportServiceProvider implements ContainerServiceProviderInterface
                     __('CSV file'),
                 ]
             );
+
+        $services
+            ->set(CustomizedRdfImporter::class, CustomizedRdfImporter::class)
+            ->public();
     }
 }
